+++ src/components/spartan/STextArea/index.ts
@@ -1,0 +1,3 @@
+export { default as STextArea } from './STextArea.vue';
+
+export type { TTextAreaProps } from './types';