+++ src/components/spartan/STextArea/types.ts
@@ -1,0 +1,5 @@
+export type TTextAreaProps = {
+    class?: string;
+    error?: boolean;
+    disabled?: boolean;
+};