+++ src/components/spartan/STextAreaBlock/index.ts
@@ -1,0 +1,1 @@
+export { default as STextAreaBlock } from './STextAreaBlock.vue';